cmd로 DB에 item 정보 추가.
